# `SmallAtom` optimization