@Rigidity 2024-02-03
This is interesting, do the bench times increase in cases where most of the atoms are large due to the extra branching to determine the type of the NodePtr?

I'm also concerned about returning a reference to a temporary value with `atom()`, this feels like a technically memory safe, but logically unsafe thing to do.

But otherwise I like this a lot