@arvidn 2024-02-03
> This is interesting, do the bench times increase in cases where most of the atoms are large due to the extra branching to determine the type of the NodePtr?

I think there's probably more noise in those timings than one might imagine. I just basically ran it once. But, yeah, there are a few cases where there's more work done. For example, when asking whether a node is a `small_int()`, I actually construct a small int, if it doesn't use the internal representation but would fit. This is probably not necessary.

There's also some extra work when allocating a new number. There's a check to see if it could be stored as a small int, and if so, it is.

> I'm also concerned about returning a reference to a temporary value with atom(), this feels like a technically memory safe, but logically unsafe thing to do.

Yeah, I'm not super happy about this. I think a ring buffer of 16 is more than enough, but maybe it's better to overshoot by more and make it 64 or so.

In the long run I would like to phase out this part of the API though, in favor of `visit_node()` (and the other getters)

This is still work in progress, as you can see from the TODO comments and missing tests