@richardkiss 2024-02-07
If I understand correctly, this is actually a `SmallInt`. Maybe we should call it that or `SmallAtomInt`?