@richardkiss 2024-02-07
This is some good work. The numbers in particular look fantastic. I do worry about subtle consensus changes and wonder if adding more `visit_node_as_XXX` functions would simplify and alleviate concerns that `SmallAtom` is always correctly coded whenever it's possible.