@arvidn 2024-02-07
@Rigidity pointed out that `visit_node()` could probably return the `NodeVisitor`, rather than taking a callback. Doing that sounds like an improvement to me. I'll investigate that.