@arvidn 2024-02-07
I applied the suggestion from @Rigidity and made `visit_node()` return the `NodeVisitor` rather than taking a callback function to pass it to. I also renamed the function just `node()`.

The change since last version: https://github.com/Chia-Network/clvm_rs/compare/34bbdf08f7d0c59fa3312b66aa1d3173839b6866..e90cd1279ae60da5f21953486c3a454f583ca217